remove `maki-code-index` now that we have the builtin index lua plugin
No need for `find_references` tool, it didn't do too much, grep is fine.
