+++ maki-agent/src/mcp/config.rs
@@ -9,10 +9,9 @@ use toml_edit::DocumentMut;
 use super::SEPARATOR;
 use super::error::McpError;
 use crate::tools::{
-    BASH_TOOL_NAME, BATCH_TOOL_NAME, CODE_EXECUTION_TOOL_NAME, EDIT_TOOL_NAME,
-    FIND_SYMBOL_TOOL_NAME, GLOB_TOOL_NAME, GREP_TOOL_NAME, MEMORY_TOOL_NAME, MULTIEDIT_TOOL_NAME,
-    QUESTION_TOOL_NAME, READ_TOOL_NAME, SKILL_TOOL_NAME, TASK_TOOL_NAME, TODOWRITE_TOOL_NAME,
-    WRITE_TOOL_NAME,
+    BASH_TOOL_NAME, BATCH_TOOL_NAME, CODE_EXECUTION_TOOL_NAME, EDIT_TOOL_NAME, GLOB_TOOL_NAME,
+    GREP_TOOL_NAME, MEMORY_TOOL_NAME, MULTIEDIT_TOOL_NAME, QUESTION_TOOL_NAME, READ_TOOL_NAME,
+    SKILL_TOOL_NAME, TASK_TOOL_NAME, TODOWRITE_TOOL_NAME, WRITE_TOOL_NAME,
 };
 use maki_config::{PROJECT_CONFIG_FILE, global_config_path};
 
@@ -27,7 +26,6 @@ const BUILTIN_TOOL_NAMES: &[&str] = &[
     MULTIEDIT_TOOL_NAME,
     GLOB_TOOL_NAME,
     GREP_TOOL_NAME,
-    FIND_SYMBOL_TOOL_NAME,
     QUESTION_TOOL_NAME,
     TODOWRITE_TOOL_NAME,
     MEMORY_TOOL_NAME,
