--- maki-agent/src/tools/find_symbol.md
@@ -1,6 +1,0 @@
-Find all references to a symbol across a project. Scope-aware: locals search enclosing function, private items search the file.
-
-- Supports Rust, C, and C++.
-- No type info: may include false positives for common names.
-- Returns file:line:col, kind (def/call/read/write/type_ref/field_ref), and context.
-- To find all references to a symbol, **prefer find_symbol over grep**!
