+++ maki-agent/src/tools/grep.md
@@ -5,4 +5,3 @@ Search file contents using regex.
 - Prefer speculative parallel searches over sequential rounds of glob+grep.
 - Do NOT wrap the pattern in quotes. Do NOT double-escape (e.g. `\[` not `\\[`).
 - Multi-line matching is auto-enabled when the pattern contains `\n`, `(?s)`, or `(?m)`.
-- To find all references to a symbol, prefer find_symbol over grep.
