+++ maki-agent/src/tools/task.rs
@@ -241,7 +241,6 @@ mod tests {
             (super::super::READ_TOOL_NAME, all),
             (super::super::GLOB_TOOL_NAME, all),
             (super::super::GREP_TOOL_NAME, all),
-            (super::super::FIND_SYMBOL_TOOL_NAME, all),
             (super::super::WRITE_TOOL_NAME, MAIN | GEN | INT),
             (super::super::EDIT_TOOL_NAME, MAIN | GEN | INT),
             (super::super::MULTIEDIT_TOOL_NAME, MAIN | GEN | INT),
