--- maki-code-index/Cargo.toml
@@ -1,69 +1,0 @@
-[package]
-name = "maki-code-index"
-version.workspace = true
-edition.workspace = true
-
-[features]
-default = [
-    "lang-bash",
-    "lang-c",
-    "lang-c-sharp",
-    "lang-cpp",
-    "lang-elixir",
-    "lang-go",
-    "lang-java",
-    "lang-kotlin",
-    "lang-lua",
-    "lang-markdown",
-    "lang-php",
-    "lang-python",
-    "lang-ruby",
-    "lang-rust",
-    "lang-scala",
-    "lang-swift",
-    "lang-typescript",
-]
-lang-bash = ["dep:tree-sitter-bash"]
-lang-c = ["dep:tree-sitter-c"]
-lang-c-sharp = ["dep:tree-sitter-c-sharp"]
-lang-cpp = ["dep:tree-sitter-cpp", "dep:tree-sitter-c"]
-lang-elixir = ["dep:tree-sitter-elixir"]
-lang-go = ["dep:tree-sitter-go"]
-lang-java = ["dep:tree-sitter-java"]
-lang-kotlin = ["dep:tree-sitter-kotlin-ng"]
-lang-lua = ["dep:tree-sitter-lua"]
-lang-markdown = ["dep:tree-sitter-md"]
-lang-php = ["dep:tree-sitter-php"]
-lang-python = ["dep:tree-sitter-python"]
-lang-ruby = ["dep:tree-sitter-ruby"]
-lang-rust = ["dep:tree-sitter-rust"]
-lang-scala = ["dep:tree-sitter-scala"]
-lang-swift = ["dep:tree-sitter-swift"]
-lang-typescript = ["dep:tree-sitter-typescript", "dep:tree-sitter-javascript"]
-
-[dependencies]
-ignore = { workspace = true }
-memchr = { workspace = true }
-thiserror = { workspace = true }
-tree-sitter = "0.26"
-tree-sitter-bash = { version = "0.25", optional = true }
-tree-sitter-c = { version = "0.24", optional = true }
-tree-sitter-c-sharp = { version = "0.23", optional = true }
-tree-sitter-cpp = { version = "0.23", optional = true }
-tree-sitter-elixir = { version = "0.3.5", optional = true }
-tree-sitter-go = { version = "0.23", optional = true }
-tree-sitter-java = { version = "0.23", optional = true }
-tree-sitter-javascript = { version = "0.23", optional = true }
-tree-sitter-kotlin-ng = { version = "1.1", optional = true }
-tree-sitter-lua = { version = "0.5", optional = true }
-tree-sitter-md = { version = "0.5.3", optional = true }
-tree-sitter-php = { version = "0.24", optional = true }
-tree-sitter-python = { version = "0.23", optional = true }
-tree-sitter-ruby = { version = "0.23", optional = true }
-tree-sitter-rust = { version = "0.24", optional = true }
-tree-sitter-scala = { version = "0.25", optional = true }
-tree-sitter-swift = { version = "0.7", optional = true }
-tree-sitter-typescript = { version = "0.23", optional = true }
-
-[dev-dependencies]
-test-case = { workspace = true }
