+++ maki-docgen/src/gen_tools.rs
@@ -15,7 +15,6 @@ const SECTIONS: &[(&str, &[&str])] = &[
             "multiedit",
             "glob",
             "grep",
-            "find_symbol",
             "index",
         ],
     ),
@@ -346,14 +345,6 @@ pub fn generate() -> String {
             }
             writeln!(out).unwrap();
             writeln!(out, "{summary}").unwrap();
-            if *name == "find_symbol" {
-                writeln!(out).unwrap();
-                writeln!(
-                    out,
-                    "> **Experimental.** Opt in with `experimental_find_symbol = true` in your config."
-                )
-                .unwrap();
-            }
             writeln!(out).unwrap();
             write_param_table(&mut out, &params);
         }
