+++ maki-lua/src/language.rs
@@ -1,0 +1,96 @@
+use tree_sitter::Language as TsLanguage;
+
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+pub enum Language {
+    Rust,
+    Python,
+    TypeScript,
+    JavaScript,
+    Go,
+    Java,
+    C,
+    Cpp,
+    CSharp,
+    Ruby,
+    Php,
+    Swift,
+    Kotlin,
+    Scala,
+    Bash,
+    Lua,
+    Elixir,
+    Markdown,
+}
+
+impl Language {
+    pub fn from_name(name: &str) -> Option<Self> {
+        match name {
+            "rust" => Some(Self::Rust),
+            "python" => Some(Self::Python),
+            "typescript" => Some(Self::TypeScript),
+            "javascript" => Some(Self::JavaScript),
+            "go" => Some(Self::Go),
+            "java" => Some(Self::Java),
+            "c" => Some(Self::C),
+            "cpp" => Some(Self::Cpp),
+            "c_sharp" => Some(Self::CSharp),
+            "ruby" => Some(Self::Ruby),
+            "php" => Some(Self::Php),
+            "swift" => Some(Self::Swift),
+            "kotlin" => Some(Self::Kotlin),
+            "scala" => Some(Self::Scala),
+            "bash" => Some(Self::Bash),
+            "lua" => Some(Self::Lua),
+            "elixir" => Some(Self::Elixir),
+            "markdown" => Some(Self::Markdown),
+            _ => None,
+        }
+    }
+
+    pub fn from_extension(ext: &str) -> Option<Self> {
+        match ext {
+            "rs" => Some(Self::Rust),
+            "py" | "pyi" => Some(Self::Python),
+            "ts" | "tsx" => Some(Self::TypeScript),
+            "js" | "jsx" | "mjs" | "cjs" => Some(Self::JavaScript),
+            "go" => Some(Self::Go),
+            "java" => Some(Self::Java),
+            "c" | "h" => Some(Self::C),
+            "cpp" | "cc" | "cxx" | "hpp" | "hxx" | "hh" => Some(Self::Cpp),
+            "cs" => Some(Self::CSharp),
+            "rb" | "rake" | "gemspec" => Some(Self::Ruby),
+            "php" => Some(Self::Php),
+            "swift" => Some(Self::Swift),
+            "kt" | "kts" => Some(Self::Kotlin),
+            "scala" | "sc" => Some(Self::Scala),
+            "sh" | "bash" | "zsh" => Some(Self::Bash),
+            "lua" => Some(Self::Lua),
+            "ex" | "exs" => Some(Self::Elixir),
+            "md" | "markdown" => Some(Self::Markdown),
+            _ => None,
+        }
+    }
+
+    pub fn ts_language(&self) -> TsLanguage {
+        match self {
+            Self::Rust => tree_sitter_rust::LANGUAGE.into(),
+            Self::Python => tree_sitter_python::LANGUAGE.into(),
+            Self::TypeScript => tree_sitter_typescript::LANGUAGE_TYPESCRIPT.into(),
+            Self::JavaScript => tree_sitter_javascript::LANGUAGE.into(),
+            Self::Go => tree_sitter_go::LANGUAGE.into(),
+            Self::Java => tree_sitter_java::LANGUAGE.into(),
+            Self::C => tree_sitter_c::LANGUAGE.into(),
+            Self::Cpp => tree_sitter_cpp::LANGUAGE.into(),
+            Self::CSharp => tree_sitter_c_sharp::LANGUAGE.into(),
+            Self::Ruby => tree_sitter_ruby::LANGUAGE.into(),
+            Self::Php => tree_sitter_php::LANGUAGE_PHP.into(),
+            Self::Swift => tree_sitter_swift::LANGUAGE.into(),
+            Self::Kotlin => tree_sitter_kotlin_ng::LANGUAGE.into(),
+            Self::Scala => tree_sitter_scala::LANGUAGE.into(),
+            Self::Bash => tree_sitter_bash::LANGUAGE.into(),
+            Self::Lua => tree_sitter_lua::LANGUAGE.into(),
+            Self::Elixir => tree_sitter_elixir::LANGUAGE.into(),
+            Self::Markdown => tree_sitter_md::LANGUAGE.into(),
+        }
+    }
+}
