+++ maki-ui/src/components/render_hints.rs
@@ -1,10 +1,9 @@
 use crate::markdown::Keep;
 use maki_agent::RawRenderHints;
 use maki_agent::tools::{
-    BASH_TOOL_NAME, BATCH_TOOL_NAME, CODE_EXECUTION_TOOL_NAME, EDIT_TOOL_NAME,
-    FIND_SYMBOL_TOOL_NAME, GLOB_TOOL_NAME, GREP_TOOL_NAME, MEMORY_TOOL_NAME, MULTIEDIT_TOOL_NAME,
-    QUESTION_TOOL_NAME, READ_TOOL_NAME, SKILL_TOOL_NAME, TASK_TOOL_NAME, TODOWRITE_TOOL_NAME,
-    WRITE_TOOL_NAME,
+    BASH_TOOL_NAME, BATCH_TOOL_NAME, CODE_EXECUTION_TOOL_NAME, EDIT_TOOL_NAME, GLOB_TOOL_NAME,
+    GREP_TOOL_NAME, MEMORY_TOOL_NAME, MULTIEDIT_TOOL_NAME, QUESTION_TOOL_NAME, READ_TOOL_NAME,
+    SKILL_TOOL_NAME, TASK_TOOL_NAME, TODOWRITE_TOOL_NAME, WRITE_TOOL_NAME,
 };
 use std::collections::HashMap;
 use std::sync::Arc;
@@ -128,7 +127,6 @@ const DEFAULT_HINTS: &[(&str, ToolRenderHints)] = &[
     hint!(EDIT_TOOL_NAME, header_style: HeaderStyle::Path),
     hint!(MULTIEDIT_TOOL_NAME, header_style: HeaderStyle::Path),
     hint!(MEMORY_TOOL_NAME, header_style: HeaderStyle::Path),
-    hint!(FIND_SYMBOL_TOOL_NAME),
     hint!(TODOWRITE_TOOL_NAME),
     hint!(QUESTION_TOOL_NAME),
     hint!(BATCH_TOOL_NAME),
